move unused containerfilemanager to internal pkg
