+++ cmd/monitor/monitor/main.go
@@ -29,7 +29,6 @@ import (
 	_ "github.com/erda-project/erda-proto-go/core/messenger/notifygroup/client"
 	_ "github.com/erda-project/erda-proto-go/core/org/client"
 	_ "github.com/erda-project/erda-proto-go/core/pipeline/cron/client"
-	_ "github.com/erda-project/erda/internal/core/file-manager"
 	_ "github.com/erda-project/erda/internal/tools/monitor/apm/runtime"
 	_ "github.com/erda-project/erda/internal/tools/monitor/apm/topology"
 	_ "github.com/erda-project/erda/internal/tools/monitor/core/alert/alert-apis"
