+++ internal/pkg/containerfilemanager/upload.html
